# master/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/cron/task/kill": {
            "post": {
                "description": "kill task by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "kill task by name",
                "parameters": [
                    {
                        "description": "request params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.KillInputTask"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/api/cron/task/log/{name}": {
            "get": {
                "description": "get log list by task name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "get log list by task name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "request params",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/master.TaskLog"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cron/tasks": {
            "get": {
                "description": "get task list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "get task list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Task"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "save task to db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "save task to db",
                "parameters": [
                    {
                        "description": "request params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Task"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cron/tasks/{name}": {
            "delete": {
                "description": "delete task by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "delete task by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "request params",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Task"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/cron/workers": {
            "get": {
                "description": "get worker list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "get worker list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "any"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "master.TaskLog": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "err": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "plan_time": {
                    "type": "string"
                },
                "schedule_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
        "model.KillInputTask": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Task": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "cron_expr": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}